requirements.txt
Django>=4.2
djangorestframework
psycopg2-binary
pandas
celery[redis]
redis
python-dotenv
openpyxl
gunicorn
